fix: post init logic not launch
